Fix hover over list shadow bug
This was caused by the target offset being reset to 0 when hovering over
the ul element. Using a CSS grid when would only add the offset of the
grid divs.

This commit adds the offset of the ul element if the target is withen
the list.

This problem may be encountered later if more elements are added within
the individual grid element. It is not worth comming up with a
encompassing solution until we run into this again.

A more permanment fix may be to check if the target is within a element
that is nested within another element, then add the offset.