Only add the offset of <ul> when target is inside the element
There were a few pixels above the list that were the actual <ul>
element. We would add the offset of the list on top of the offset of the
target which would result in a similar bug to the hover bug.

By only adding the list offset when the target is inside the <ul> but
not the <ul> we fix this.